fastapi-template/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pymongo
pydantic
python-jose[cryptography]
passlib[bcrypt]
firebase-admin
requests
python-dotenv
pandas
numpy
scikit-learn
joblib
geojson
eth-account
web3
motor
loguru